(original rev 2412) merge from trunk into staging
Merged revisions 2392-2406,2409 via svnmerge from
file:///gaba/svn/netgpu/omni/trunk

........
  r2392 |  | 2010-08-10 17:00:03 -0400 (Tue, 10 Aug 2010) | 2 lines

  add make clean omni option
........
  r2393 |  | 2010-08-10 17:34:04 -0400 (Tue, 10 Aug 2010) | 2 lines

  revert floating toolbar change
........
  r2394 |  | 2010-08-10 17:41:05 -0400 (Tue, 10 Aug 2010) | 2 lines

  don't die on resizing
........
  r2395 |  | 2010-08-11 15:11:14 -0400 (Wed, 11 Aug 2010) | 2 lines

  don't display multiple toolbars
........
  r2396 |  | 2010-08-11 15:55:31 -0400 (Wed, 11 Aug 2010) | 2 lines

  start of refactor of segmentlistsbysize
........
  r2397 |  | 2010-08-11 16:24:33 -0400 (Wed, 11 Aug 2010) | 2 lines

  refactor recent list into one place
........
  r2398 |  | 2010-08-11 16:28:45 -0400 (Wed, 11 Aug 2010) | 2 lines

  small cleanup
........
  r2399 | mwimer | 2010-08-11 16:56:33 -0400 (Wed, 11 Aug 2010) | 2 lines

  fast paint
........
  r2400 | mwimer | 2010-08-11 18:44:04 -0400 (Wed, 11 Aug 2010) | 2 lines

  segment granularity bounding boxes for centering
........
  r2401 | mwimer | 2010-08-11 18:47:46 -0400 (Wed, 11 Aug 2010) | 2 lines

  based on empty bounds not version
........
  r2402 | mwimer | 2010-08-11 18:52:43 -0400 (Wed, 11 Aug 2010) | 2 lines

  speed hack...hahah
........
  r2403 | mwimer | 2010-08-11 18:55:05 -0400 (Wed, 11 Aug 2010) | 2 lines

  ....stuff...2 * .9
........
  r2404 |  | 2010-08-11 18:58:43 -0400 (Wed, 11 Aug 2010) | 2 lines

  time the graph load
........
  r2405 |  | 2010-08-11 19:00:57 -0400 (Wed, 11 Aug 2010) | 2 lines

  don't buffer tiles
........
  r2406 | mwimer | 2010-08-11 19:16:52 -0400 (Wed, 11 Aug 2010) | 2 lines

  100
........
  r2409 |  | 2010-08-12 01:45:06 -0400 (Thu, 12 Aug 2010) | 2 lines

  switch over to zi threads for computing which meshes to draw in 3d
........




git-svn-id: file:///tmp/current/omni/branches/staging@2169 2ffc8166-1744-45da-a82a-d6ea10f7896b

+++ omni/src/segment/lowLevel/omSegmentListBySize.h
@@ -28,7 +28,7 @@ class OmSegmentListBySize
 	OmSegIDsListWithPage * getAPageWorthOfSegmentIDs( const unsigned int, const int, const OmSegID);
 	quint64 getSegmentSize( OmSegment * seg );
 
-	quint32 size();
+	size_t size();
 	void dump();
 	void clear();
 
@@ -42,7 +42,7 @@ class OmSegmentListBySize
 	struct OmSegSize {
 		OmSegID segID;
 		quint64 segSize;
-		OmSegSize( const OmSegID segID_, const quint64 segSize_ ) 
+		OmSegSize( const OmSegID segID_, const quint64 segSize_ )
 			: segID(segID_), segSize(segSize_){}
 	};
